Added a switch for short output: -c
For usage in scripts I do not want to parse the output and filter out
the usefull infomration or clutter the line with redirects to /dev/null.
Therefore I added a option, that removes every output, except whith the
-b-switch, where the battery-percentage (or -1 for loading) are printed.

While doing so, the default changed a bit:
Only when the -h-swich is present, we output name and author of this
programm.